Do not use unsafe block.
The encode module contained unsafe code. While tiny and easily
validated, the presence of an unsafe block has the potential to rub
people the wrong way, particularly if they're doing any kind of
automated validation of their codebase (think: `$ rg unsafe`), so it
probably behooves us to just not have this here--particularly given the
specious performance improvements associated with the omission of this
bounds check.