Beta (#2)
* Clippy lints

* Do not use unsafe block.

The encode module contained unsafe code. While tiny and easily
validated, the presence of an unsafe block has the potential to rub
people the wrong way, particularly if they're doing any kind of
automated validation of their codebase (think: `$ rg unsafe`), so it
probably behooves us to just not have this here--particularly given the
specious performance improvements associated with the omission of this
bounds check.

* Cargo fmt

* Add docs

* Update readme

* Fuzz error: wrapping multiplication.

* Fuzz error: check digit

Fuzz testing revealed that inputing the letter U would cause the world
to burn. This is because, unlike i, l, and o, u was not filtered before
we hit the letter conversion section. U is now filtered.

* Add fuzzing

* Hide bitly iterator

* Beta-ify

* Add specific check digit error.

* Fixup readme

* fixup docs

* Expose error type

* Fixup docs again