Collect encoded fits to ASCII buffer before String
A `char` may be encoded into anywhere from 1 to 4 bytes when it comes to
UTF-8. Because all input characters are ASCII, we know that they all use
only one byte when encoded as UTF-8. This should result in only a single
allocation when creating the string, allowing for better performance.