Implement left to right encoding.
Left to right encoding also gives us the opportunity to encode to a
buffer in one pass, without an intermediate buffer used in reversing the
values as they come out. We've taken advantage of this here with an
experimental-ish API accepting a custom `Write` trait.

Concerns:

1. I'm not convinced that it's worthwhile to introduce generics into
such a simple library. Ideally, this should be very lightweight, and
sticking in code that needs to be specialized seems like a bad plan.
2. I'm also not convinced we have the right API here: I'm offering a
write impl for both `Vec<u8>` and `String`, but the UTF8 validation
behavior of String has the potential to cause the "fast" #encode_into
variant of the function slower than the other one.