Write::write() is unsafe
For a while, we've had it so that Write::write() will silently modify
the values being written to the output, but that has never felt right.
To deal with the safety issues presented by the fact that we don't
bother to check whether a u8 is actually an ASCII byte, we are now just
marking the darned function as unsafe.